Update patch set 1
Patch Set 1: Code-Review+2

In signal handler we can wait for another external thread to enter Go and create an extram. Which in turn can wait for stoptheworld to finish. Which in turn can wait for all other threads to reach safe points. Looks like a deadlock waiting to happen. But this patch does not make things worse, so +2.

It's worrying that we started to drop SIGPROFs on the floor again in sigtrampgo and in sigtrampnog. Skewed profiles are of no use and waste developers time. I almost started to trust Go's pprof...

Patch-set: 1
Reviewer: Gerrit User 5400 <5400@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+2