Update patch set 2
Patch Set 2:

Brad and Adam seem to both be leaning in the "every little bit helps" direction, and I lean that way too for this (fairly minimal) change. I checked the timing on the new ConstantTimeSum, and it raises BenchmarkHash8Bytes from 177 ns/op to 433 ns/op on my system, which is probably (?) too slow to make just the default implementation of Sum instead of exposing new (hidden) methods. For now this seems fine. If we can somehow make ConstantTimeSum less of a hit, it seems worth considering making it the default Sum.

I edited the patch to take out the deletion of the warning. (That is, I restored the warning.)

Patch-set: 2