Update patch set 1
Patch Set 1:

> In signal handler we can wait for another external thread to enter Go and create an extram. Which in turn can wait for stoptheworld to finish. Which in turn can wait for all other threads to reach safe points. Looks like a deadlock waiting to happen. But this patch does not make things worse, so +2.

I'm not sure there is a deadlock.  I agree that the thread on which the signal occurred might have to wait for stoptheworld to finish.  That could happen if a Go function is called from C code running on a non-Go thread, and that thread is preempted between the call to needm in cgocallback_gofunc and the call to newextram in cgocallbackg1.  However, I don't see why the stoptheworld would have to wait for the thread on which the signal occurred before finishing.  For this code path, the thread on which the signal occurred is a non-Go thread, so stoptheworld doesn't know anything about it.


> It's worrying that we started to drop SIGPROFs on the floor again in sigtrampgo and in sigtrampnog. Skewed profiles are of no use and waste developers time. I almost started to trust Go's pprof...

We haven't started to drop them again.  We've always dropped SIGPROF signals that occur on a non-Go thread.  That is because in general we have no way to get a stack trace to record them.  That said, note that on amd64 GNU/Linux, if you call runtime.SetCgoTraceback, SIGPROF on a non-Go thread is recorded correctly, via sigprofNonGo.  It would be straightforward to extend that support to other GOOS/GOARCH combinations.

Patch-set: 1