Update patch set 1
Patch Set 1:

>  > > It's worrying that we started to drop SIGPROFs on the floor again
>  > in sigtrampgo and in sigtrampnog. Skewed profiles are of no use and
>  > waste developers time. I almost started to trust Go's pprof...
>  > 
>  > We haven't started to drop them again.  We've always dropped
>  > SIGPROF signals that occur on a non-Go thread.  That is because in
>  > general we have no way to get a stack trace to record them.  That
>  > said, note that on amd64 GNU/Linux, if you call runtime.SetCgoTraceback,
>  > SIGPROF on a non-Go thread is recorded correctly, via sigprofNonGo.
>  >  It would be straightforward to extend that support to other
>  > GOOS/GOARCH combinations.
> 
> 
> We did not record a stack trace, we recorded just the current PC for these cases. It is not fully informative, but at least it does not skew the rest of the profile (i.e. you are still sure that the top function in the profile is correct and you don't spend time optimizing something that is not significant).

That is a good idea, but I looked back through several releases and I can't find any release in which we actually did that.


> > note that on amd64 GNU/Linux, if you call runtime.SetCgoTraceback, SIGPROF on a non-Go thread is recorded correctly, via sigprofNonGo
> 
> It doesn't, see sigtrampgo/sigprofCallersUse

On amd64 GNU/Linux, in a program that uses cgo and calls runtime.SetCgoTraceback, we normally do not get to sigtrampgo if a SIGPROF occurs on a non-Go thread.  In a program that uses cgo, the signal handler is cgoSigtramp in sys_linux_amd64.s.

It's true that if we get a SIGPROF signal we will lock sigprofCallersUse while we collect a stack trace, and if a second SIGPROF signal occurs during that time we will drop the second signal.  We could fix that to, as you suggest, record just the current PC.

Patch-set: 1