Update patch set 1
Patch Set 1:

(1 comment)

Can we add a test that we've covered all the ops that need it?
Some sort of assertion in AddrScratch.  Maybe we'd die anyway because s.ScratchFpMem would be nil, but better to get an error message "Op ? asked for scratch space but was not declared to need any" than a segfault.

Patch-set: 1