Update patch set 1
Patch Set 1:

Apologies for taking so long to get to this CL. It's very well done. The only thing missing is documentation for the new behavior.

That said, it looks like CL 20686 made it into Go 1.7 with slightly different semantics for %b on []byte, namely, analogous to %b on []int instead of analogous to %x on []byte. Since that's been released we can't change the semantics very easily now, but after some thought I think that CL's interpretation is probably the right one. Although it's common to think of a hex dump of a string as analogous to a string, it's rarer to think of a binary dump of a string as analogous to a string, and the boundaries are more important given the longer representations of individual elements.

Sorry again for the long delay, and thanks for the CL.

Patch-set: 1
Reviewer: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0