Update patch set 1
Patch Set 1:

> I should add: there is a possible alternative way to make this
 > faster. The compiler is smart enough now to substitute nice loads
 > where possible when it sees references to binary.LittleEndian.Uint32
 > etc. So one possible option is:
 > 
 > var v uint32
 > var i int
 > for i = 0; i+4 <= len(x); i += 4 {
 > v |= binary.LittleEndian.Uint32(x[i:i+4]) ^ binary.LittleEndian.Uint32(y[i:i+4])
 > }
 > for ; i < len(x); i++ {
 > v |= uint32(x[i] ^ y[i])
 > }
 > return ConstantTimeEq(int32(v), 0)
 > 
 > If that makes a significant speed difference then - because it is
 > so much simpler than the current CL - it seems fine to check in. At
 > the least though you need to put in benchmarks for small and large
 > inputs in the CL description.
 > 
 > Thanks.

Russ: the original rationale for this change was slowness with header processing in gRPC.  The API in Go (and in Java too) is based on strings.  Checking if a string is present in the HPACK table is needed to see if an index should be used in the headers.  Since these headers can be sensitive, we use the constant time code here.

These headers are sent a lot, and sometimes are big.  This is the case for OAuth tokens, and JWTs, which can be 50-200 bytes long.  They are sent with every RPC (http request) so the HPACK code starts showing up on profiling.  I believe Brad noticed something similar.

I am not sure where I can get realistic benchmarks, but I can probably look at some in the gRPC Go code.  Would you accept those?

Patch-set: 1
Reviewer: Gerrit User 12225 <12225@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0