runtime, cmd/trace: annotate different mark worker types
Currently mark workers are shown in the trace as regular goroutines
labeled "runtime.gcBgMarkWorker". That's somewhat unhelpful to an end
user because of the opaque label and particularly unhelpful to runtime
developers because it doesn't distinguish the different types of mark
workers.

Fix this by introducing a variant of the GoStart event called
GCWorkerStart that records the worker mode and modifying the trace
viewer to display these as "GC (<mode>)".

Change-Id: Id7b9c0536508430c661ffb9e40e436f3901ca121