runtime, cmd/trace: annotate different mark worker types
DO NOT SUBMIT. Add Go 1.7 trace to internal/trace/testdata.

Currently mark workers are shown in the trace as regular goroutines
labeled "runtime.gcBgMarkWorker". That's somewhat unhelpful to an end
user because of the opaque label and particularly unhelpful to runtime
developers because it doesn't distinguish the different types of mark
workers.

Fix this by introducing a variant of the GoStart event called
GoStartLabel that lets the runtime indicate a label for a goroutine
execution span and using this to label mark worker executions as "GC
(<mode>)" in the trace viewer.

Change-Id: Id7b9c0536508430c661ffb9e40e436f3901ca121