Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> I don't see the benefit of this. I like the old way, not duplicating the constant, and leaving open the possibility to turn the constant into a function call or something.

I can change it to do that again. The benefit was that I moved the if statement under the else of the previous if statement because if the delay was just set to 5 milliseconds, how could it become greater than a second?

Patch-set: 1
Reviewer: Gerrit User 9620 <9620@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0