cmd/compile,runtime: redo how map assignments work
To compile:
  m[k] = v
instead of:
  mapassign(maptype, m, &k, &v), do
do:
  *mapassign(maptype, m, &k) = v

mapassign returns a pointer to the value slot in the map.  It is just
like mapaccess except that it will allocate a new slot if k is not
already present in the map.

This CL is in preparation for doing operations like m[k] += v with only
a single runtime call.

DO NOT REVIEW
DO NOT SUBMIT

Change-Id: Ia435f032090a2ed905dac9234e693972fe8c2dc5