Update patch set 1
Patch Set 1:

Yes, I have verified that it disassembles correctly for all CR values (using a hacked tool I have to do that).  I don't think there are any testcases yet that do extensive verification that the assembler's encoding of the generated instruction is correct, except for the testcase you mention but it has only 4 or 5 instructions.  There is another asm testcase with more instructions being tested but that only verifies that the assembler doesn't flag any errors, it doesn't try to figure out if the generated instruction is actually correct.
Once we get the disassembler working then we can add an example to a disassembly test to verify this.

Patch-set: 1