testing/quick: omit internal read-only fields
Omit internal read only fields, otherwise the function "sizedValue" raise panic, when create a struct with internal fields.
Added checking the implementation of the Generator interface by type to which is accessed
through a pointer, otherwise happens implicitly receive variable and access to it, as well
as a pointer is nil, then panic. After checking is called a generator directly for the type,
and then creates a pointer to it.

Change-Id: I6c5a06460646d66928b92ba4278bae8dd6df9051