Update patch set 1
Patch Set 1:

This is pretty light on documentation, but so is the rest of the package. If you want me to take the occasion to document the rest of the methods let me know. (In particular Val says nothing about concurrent access requirements.)

Patch-set: 1
Reviewer: Gerrit User 11715 <11715@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0