Update patch set 2
Patch Set 2:

(10 comments)

Thanks for the review @adg. I've updated the example with your suggestions. However I still used .String() because in the Unmarshal code we print out a slice and without a .String() method, we wouldn't have output to show them. If you think the example using ints as the on wire representation would suffice, please let me know and I can change the example to use int and get rid of the .String(), but I think without it, the example becomes loses output readability.

Patch-set: 2