Update patch set 4
Patch Set 4: Run-TryBot+1 Code-Review+2

I'm in two minds about this: in general if you pass a nil pointer into something that's not expecting it, it'll panic and that's fine. The reason that I pause here is that Marshal is using reflection and might be walking very complex data structures. Thus it's more plausible that returning an error makes sense.

Patch-set: 4
Label: Code-Review=+2
Label: Run-TryBot=+1