Update patch set 1
Patch Set 1:

> I find it difficult to consider "one or more" as clear user facing
 > documentation with respect to zero byte behaviour.

It is an explicit goal NOT to just go around writing "This implementation does X on 0-length reads or 0-length writes". The point is that the behavior follows from the general definition; it's not just a random choice added to the API, much as some people might think that.

Patch-set: 1