Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > I find it difficult to consider "one or more" as clear user facing
>  > documentation with respect to zero byte behaviour.
> 
> It is an explicit goal NOT to just go around writing "This implementation does X on 0-length reads or 0-length writes". The point is that the behavior follows from the general definition; it's not just a random choice added to the API, much as some people might think that.

Zero is a special case whether you want it to be or not.

I think it's totally reasonable that somebody could think that a io.Writer.Write of zero bytes is a no-op. Or that an io.Reader.Read of zero bytes is a no-op.

As it stands, every implementation decides (or leaves it undefined).

This CL is nice in that it at least defines it (implicitly), but the general problem if zero-byte reads & writes to io.Readers/Writers being minefields remains.

Patch-set: 1