Update patch set 1
Patch Set 1:

I was not trying to mandate new behavior, but it looked to me like C.malloc already did this. I see:

extern void runtime_throw(const char *);
void *_cgoPREFIX_Cfunc__CMalloc(size_t n) {
        void *p = malloc(n);
        if(p == NULL && n == 0)
                p = malloc(1);
        if(p == NULL)
                runtime_throw("runtime: C malloc failed");
        return p;
}

Is that the wrong one? I'm fine with 31768 but I was trying to just document, not change, the current behavior. Or is it only true sometimes?

Patch-set: 1