Update patch set 6
Patch Set 6:

I wasn't sure how this latest version could even work, so I patched it in locally to understand how it was passing, but...

It's not passing. It deadlocks, as I thought.

Oh, and now I see the trybots weren't run on the last version.

Are you not testing these revisions locally before uploading them?

Us running the trybots shouldn't be the first time new code gets tested.

Patch-set: 6