runtime: fix race shadow mapping for data
The code was assuming that racedatastart was page aligned. It may not be.
Not sure what rounding __tsan_map_shadow does, but clearly some:
maybe it rounds the overall size down to page alignment, or maybe it
rounds the start address down without adjusting the size.
Either way, arrange for aligned arguments, and OS X (which has an
unaligned racedatastart) is happy again.

Fixes #17065.

Change-Id: I79efbf77c8b88879a104f27b21bb7ff02d027ca3