Update patch set 2
Patch Set 2:

Right now if you run

pprof --pdf binary profile

it writes a file instead of writing to the terminal, yes? 

After this CL it seems like it will write the PDF to the terminal. That's the opposite of what the CL seems to be trying to do.

Patch-set: 2