runtime: include panic/throw logs in core dumps
When a Go program crashes with GOTRACEBACK=crash, the OS creates a
core dump. Include the text-formatted output of the cause of that
crash in the core dump.

Output printed by the runtime before crashing is maintained in a
circular buffer, in particular to allow access to messages that may
be printed immediately before calling runtime.throw.

As the runtime crashes and prints stack traces, store a prefix of the
output in a separate buffer. The prefix should be enough to see the
full stack of the goroutine that encountered the fatal condition.

Fixes #16893

DO NOT SUBMIT

There are no tests covering the generation of core dumps; this change
has not added any.

This adds (reentrant) locking to runtime.gwrite, which may have an
undesired performance impact.

Change-Id: Ia2463be3c12429354d290bdec5f3c8d565d1a2c3