Update patch set 1
Patch Set 1:

> I know that it's not all that much space, but it still feels weird
 > to me to allocate these large arrays in the data section that are
 > essentially never used.

Is it correct that arrays in the data section increase the size of the executable on disk, and is that the root of the weird feeling?

Can I cause the arrays to be "allocated" in the bss section while still being arrays, and will that address your concern?

I think (though have not demonstrated) that using arrays here rather than mallocgc-allocated slices will make it possible to extract the crash message from a core dump that's being piped into a reporting program rather than written to disk. I think that being able to extract the message without writing the core file to disk is an important property to target.

I've also heard of solutions to this problem that involve creating a new and separate ELF section to hold the panic output, though it seems like a much more invasive change.

Patch-set: 1